+++ internal/rubyserver/balancer/balancer_test.go
@@ -2,6 +2,7 @@ package balancer
 
 import (
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 )
@@ -12,43 +13,62 @@ func TestRemovals(t *testing.T) {
 		{add: "bar"},
 		{add: "qux"},
 		{remove: "bar"},
+		{add: "baz"},
 		{remove: "foo"},
 	}
+	numAddr := 3
+	removeDelay := 1 * time.Millisecond
+	ConfigureBuilder(numAddr, removeDelay)
 
 	testCases := []struct {
 		desc      string
 		actions   []action
 		lastFails bool
+		delay     time.Duration
 	}{
 		{
 			desc:    "add then remove",
 			actions: okActions,
+			delay:   2 * removeDelay,
 		},
 		{
-			desc:      "remove last address",
+			desc:      "add then remove but too fast",
+			actions:   okActions,
+			lastFails: true,
+			delay:     0,
+		},
+		{
+			desc:      "remove one address too many",
 			actions:   append(okActions, action{remove: "qux"}),
 			lastFails: true,
+			delay:     2 * removeDelay,
 		},
 		{
 			desc: "remove unknown address",
 			actions: []action{
 				{add: "foo"},
+				{add: "qux"},
+				{add: "baz"},
 				{remove: "bar"},
 			},
 			lastFails: true,
+			delay:     2 * removeDelay,
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.desc, func(t *testing.T) {
-			// This breaks integration with gRPC and causes a monitor goroutine leak.
-			// Not a problem for this test.
-			lbBuilder = newBuilder()
+			lbBuilder.testingRestart <- struct{}{}
+			time.Sleep(2 * removeDelay) // wait for lastRemoval in monitor goroutine to be long enough ago
 
 			for i, a := range tc.actions {
 				if a.add != "" {
 					AddAddress(a.add)
 				} else {
+					if tc.delay > 0 {
+						time.Sleep(tc.delay)
+					}
+
 					expected := true
 					if i+1 == len(tc.actions) && tc.lastFails {
 						expected = false